settings: move extensions to proper namespace
